Added support for newlines and Sequence/Mapping
Now, newlines in labels and values are converted to \\n which displays
as \n when printed.  Without this the borders would be messed up if the
objects __str__ method returns a string containing any new lines.

Also changed checks from list and dict to Sequence (except str) and
Mapping.  This way prnt can accept anything conforming to this protocol.
This is particularly important for OrderedDict as otherwise there is no
way to control the order in which dictionaries are printed.